% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaddata.R
\name{download_data}
\alias{download_data}
\title{Descarregar dades}
\usage{
download_data(comarca)
}
\arguments{
\item{comarca}{Nom de la comarca en majuscules}
}
\value{
Dataframe amb les dades de la poblacio general de la comarca corresponent
}
\description{
Descarrega les dades més recents i crea un dataframe amb les dades de la població general de la comarca
}
\examples{
bllobregatdata <- download_data("BAIX LLOBREGAT")
}
