% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketweet.R
\name{make_tweet}
\alias{make_tweet}
\title{Crear un tweet}
\usage{
make_tweet(dataframe)
}
\arguments{
\item{dataframe}{dataframe amb les dades covid de la comarca corresponent}
}
\value{
casos confirmats, incidencia acumulada
per 100.000 habitants en els ultims 7 i 14 dies i el total
acumulat de persones vacunades
}
\description{
Crea un tweet amb les dades diaries de la comarca
}
\examples{
tweet <- make_tweet(bllobregatdata)
}
