% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdata.R
\name{plot_data}
\alias{plot_data}
\title{Crear un grafic}
\usage{
plot_data(dataframe, save = FALSE)
}
\arguments{
\item{dataframe}{Dataframe amb les dades covid de la comarca corresponent}

\item{save}{Si save = TRUE, es guarda el fitxer i es retorna el path del fitxer}
}
\value{
Si save = TRUE, retorna el path de la imatge guardada. Si no, nomes es visualitza el grafic en R studio
}
\description{
Crea un grafic amb la quantitat acumulada de vacunats al llarg del 2021
}
\examples{
\dontshow{
bllobregatdata <- download_data("BAIX LLOBREGAT")
}
plot_data(bllobregatdata)
plotfile <- plot_data(bllobregatdata, save = TRUE)
}
