% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posttweet.R
\name{post_tweet}
\alias{post_tweet}
\title{Publicar un tweet}
\usage{
post_tweet(tweet, image = NULL)
}
\arguments{
\item{tweet}{text que publicar}

\item{image}{fitxer de la imatge que es vol incloure}
}
\description{
Publica un tweet
}
\examples{
\dontrun{
post_tweet("Hello")
post_tweet("Hello", image = "image.png")
}
}
