% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posttweet.R
\name{post_tweet}
\alias{post_tweet}
\title{Publicar un tweet}
\usage{
post_tweet(tweet, keysfile = "accesskeys.json", image = NULL)
}
\arguments{
\item{tweet}{text que publicar}

\item{keysfile}{Fitxer JSON que contingui les variables:
consumerKey, consumerSecret, accessToken, accessTokenSecret.
El nom del fitxer per defecte es "accesskeys.json"}

\item{image}{fitxer de la imatge que es vol incloure}
}
\description{
Publica un tweet
}
\examples{
\dontrun{
post_tweet("Hello")
post_tweet("Hello", image = "image.png")
post_tweet("Hello", keysfile = "keys.json")
}
}
